upgrade to newest boilerplate code
- first try
