+++ blog/blocks/video/video.js
@@ -1,4 +1,4 @@
-import { buildFigure } from '../../scripts/scripts.js';
+import { buildFigure } from '../../../scripts/scripts.js';
 
 export default function decorate(block) {
   if (block.classList.contains('is-loaded')) {
