+++ head.html
@@ -1,4 +1,4 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/blog/scripts/scripts.js" type="module"></script>
+<script src="/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/blog/styles/styles.css"/>
 <link rel="icon" href="data:,">
