+++ scripts/canary.js
